+++ tests/callbacks/test_progress_bar.py
@@ -1,7 +1,7 @@
 import pytest
 
 from pytorch_lightning import Trainer
-from pytorch_lightning.callbacks import ProgressBarBase, ProgressBar, ModelCheckpoint
+from pytorch_lightning.callbacks import Callback, ProgressBarBase, ProgressBar, ModelCheckpoint
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from tests.base import EvalModelTemplate
 
@@ -193,3 +193,36 @@ def test_progress_bar_progress_refresh(tmpdir, refresh_rate):
 
     trainer.test(model)
     assert progress_bar.test_batches_seen == progress_bar.total_test_batches
+
+
+@pytest.mark.parametrize(['limit_val_batches', 'expected'], [
+    pytest.param(0, 0),
+    pytest.param(5, 7),
+])
+def test_num_sanity_val_steps_progress_bar(tmpdir, limit_val_batches, expected):
+    """
+    Test val_progress_bar total with 'num_sanity_val_steps' Trainer argument.
+    """
+    class CustomCallback(Callback):
+        def __init__(self):
+            self.val_progress_bar_total = 0
+
+        def on_validation_epoch_end(self, trainer, pl_module):
+            self.val_progress_bar_total += trainer.progress_bar_callback.val_progress_bar.total
+
+    model = EvalModelTemplate()
+    cb = CustomCallback()
+
+    trainer = Trainer(
+        default_root_dir=tmpdir,
+        max_epochs=1,
+        num_sanity_val_steps=2,
+        limit_train_batches=0,
+        limit_val_batches=limit_val_batches,
+        callbacks=[cb],
+        logger=False,
+        checkpoint_callback=False,
+        early_stop_callback=False,
+    )
+    trainer.fit(model)
+    assert cb.val_progress_bar_total == expected
